Dev docker environment
Modifies the Makefile, Dockerfile & tasks.json to allow to do dev
work within a Docker container. This has the advantage of now the
only thing I need for a dev environment is Docker & a text editor
(no need for Python, virtual environments, pelican, etc).

Not sure if I want to use this or not, on my Mac this is sometimes
*slow* (googling seems to indicate that this is a Docker Mac issue
where sometimes mounted volume access is slow).  It's sporadic
though, sometimes it works flawlessly.  Very frustrating.